Fix syntastic submodule not added correctly
The issue was that the syntastic repo was cloned into bundle/ and not
actually registered as submodule (with `git submodule add ...`). It was
committed as submodule, but without being added to `.gitmodules`.

The solution is to remove the submodule from the cache (nothing changed
on disk in the submodule)
    $ git rm --cached bundle/syntastic
and then to re-add it correcty using
    $ git submodule add https://github.com/vim-syntastic/syntastic.git bundle/syntastic
followed by a commit (actually this one here).