Add a mapping to fix incorrect Vim size
When opening a session that was saved while Vim had a particular size
(number of lines and cols), opening that session inside a tmux window
makes Vim believe the terminal still has that old size. Working in Vim
then produces garbage on the screen.

The new mapping simply executes a trivial external command and redraws
the screen. This seems to make Vim query the terminal size and obtaining
the correct values.